+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
   - Transparent enhanced image output option
+  - Added check for encrypted LRIT files in ``lrit-img.py`` and ``lrit-add.py``
 
 ### Changed
   - Default key file name **(check when upgrading from an old version)**